Extracted HTTP range tests into separate integration tests.
